requirements.txt
selenium
requests
pytest
pyyaml